changed reputation field to float rather than int in users table

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -18,7 +18,7 @@ class CreateUsersTable extends Migration {
             $table->string('email')->unique();
             $table->string('password');
             $table->string('about');
-            $table->integer('reputation');
+            $table->float('reputation');
             $table->string('api_token', 64)->unique();
             $table->rememberToken();
             $table->timestamps();
